added Android support

+++ composeApp/src/commonMain/kotlin/de/ronnyporsch/train_controller/gamepad/GamepadEvent.kt
@@ -3,8 +3,8 @@ package de.ronnyporsch.train_controller.gamepad
 sealed class GamepadEvent(open val gamepadIndex: Int) {
     data class ButtonPressed(override val gamepadIndex: Int, val button: GamepadButton,) : GamepadEvent(gamepadIndex)
     data class ButtonReleased(override val gamepadIndex: Int, val button: GamepadButton) : GamepadEvent(gamepadIndex)
-    data class LeftTriggerMoved(override val gamepadIndex: Int, val value: Int) : GamepadEvent(gamepadIndex)
-    data class RightTriggerMoved(override val gamepadIndex: Int, val value: Int) : GamepadEvent(gamepadIndex)
+    data class LeftTriggerMoved(override val gamepadIndex: Int, val value: Float) : GamepadEvent(gamepadIndex)
+    data class RightTriggerMoved(override val gamepadIndex: Int, val value: Float) : GamepadEvent(gamepadIndex)
     data class LeftStickMoved(override val gamepadIndex: Int, val x: Int, val y: Int) : GamepadEvent(gamepadIndex)
     data class RightStickMoved(override val gamepadIndex: Int, val x: Int, val y: Int) : GamepadEvent(gamepadIndex)
 }
